+++ src/angular-cesium-widgets/components.ts
@@ -1,0 +1,1 @@
+export * from './directives/draggable-to-map.directive';