+++ src/angular-cesium-widgets/services.ts
@@ -1,0 +1,1 @@
+export * from './services/draggable-to-map.service';