+++ src/angular-cesium/models.ts
@@ -12,4 +12,5 @@ export { Cartesian2 } from './models/cartesian2';
 export { ContextMenuOptions } from './models/context-menu-options';
 export { BasicContextMenu } from './models/basic-context-menu';
 export { Vec3 } from './models/vec3';
+export { Vec2 } from './models/vec2';
 export { IDescription } from './models/description';