+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
 * Added customizable labels to shape editors.
+* Added `DraggableToMapDirective` directive for dragging icons from outside the map over the map.
+* Added `IconDragService` service that exposes an observable that listens to dragging with `DraggableToMapDirective`.
 
 ## 0.0.33
 ## Fixes